NamesList-17.0.0d7.txt
@michelsu wrote:
> I checked the new list. It looks good in general. Unibook loads the NamesList with no issue detected. These are the issues/findings I have so far.
> Egyptian Hieroglyph (new 44 VS):
> It all looks good. However, you should now turn on the VS table option for the extended block (not just the main block) as I have implemented VS glyphs for both blocks.
> That is:
> @~    Standardized Variation Sequences
> Instead of:
> @~    !
Done.
>
> Major:
> You missed the 2 new Tangut characters 18D1D-18D1E
Fixed.
>
> Other minor
> - The Tangut Components Supplement (18D80-18DE8) starts w/o a group header. It looks a bit odd. I had 'Components' which I think looked a bit better.
I added a subhead, but made it "Miscellaneous components" instead of
just "Components", because I think that contrasts better with the
following subheads that specify the number of strokes for each group of
components.
> - I had put the new 1F6D8 under the group header 'Map Symbols'. It is for sure a road sign (typically enclosed in a triangle). So  I would prefer to stay that way, and not just 'Miscellaneous Symbols', and as of now it creates three ugly <reserved> line under that new glyph.
Fixed.
> - By analogy/consistency with the group header for 1FAE6..1FAE7, the group header for 1FAEF should also be 'Emotion'.
O.k. Changed to that.
> - For 1FBFA (Alarm bell symbol), you don't mention it is not emoji which was part of the initial proposal text (I think). You do that for the other sets (such as 1CCFA..1CCFC)

This one I disagreed with. Unlike the other cases, which in the charts
do look like they could possibly be interpreted as stray emoji, this is
a single character that already has a subhead "Terminal graphic
character", and that, together with "SYMBOL" in the name seemed to
suffice for this one. Adding the notice in addition to the subhead in
this case seemed like it would actually detract by overkill. So I
applied the Joe Becker editorial principle. ;-)

Co-authored-by: Ken-Whistler <ken@unicode.org>