NamesList-17.0.0d8.txt (#965)
From @Ken-Whistler:
* NamesList-17.0.0d4.txt

I have done a first pass on getting a usable NamesList.txt for 17.0. […]

This is generated against my version of the 17.0 UnicodeData.txt (UnicodeData-17.0.0d1.txt […]) and an initial drop of StandardizedVariants.txt (StandardizedVariants-17.0.0d1.txt […]), and the 16.0 version of Unikemet.txt.

I manually synched the annotations against the 7th ed CD, and then updated further using Michel's latest delta names list file (which also includes most of the 18.0 anticipated repertoire). Then I also rolled in most of my action items for names list annotations for 17.0.

Michel will notice that I did some pretty extensive editing of the annotations, particularly for the various content re phonetic characters from Kirk Miller. Hopefully, the result is better, but of course we have lots of opportunity for further improvements.

This should suffice to start specific charts work for 17.0. It matches the current pipeline and is a proper subset of the content that Michel is dealing with. Let me know if you spot any egregious errors, and I'll turn around fixes as soon as I can.

* NamesList-17.0.0d5.txt

Updated again, this time after rolling in the various cuneiform annotation updates per [UTC-181-A136] (L2/24-239, Section 3.1) […]

* NamesList-17.0.0d6.txt

@eggrobin wrote:
> I compared the cuneiform changes with the chart in L2/24-239.
> You seem to be missing the annotations on U+12038 𒀸 (alias 1 aš and cross-reference to 𒐀). I don’t think this is intentional, since you have equivalent ones for U+12079 𒁹.
> Other than that the cuneiform looks good.

Yeah, I missed 12038 because it was on the other side of a pdf page boundary. Doh!

* NamesList-17.0.0d7.txt

@michelsu wrote:
> I checked the new list. It looks good in general. Unibook loads the NamesList with no issue detected. These are the issues/findings I have so far.
> Egyptian Hieroglyph (new 44 VS):
> It all looks good. However, you should now turn on the VS table option for the extended block (not just the main block) as I have implemented VS glyphs for both blocks.
> That is:
> @~    Standardized Variation Sequences
> Instead of:
> @~    !
Done.
>
> Major:
> You missed the 2 new Tangut characters 18D1D-18D1E
Fixed.
>
> Other minor
> - The Tangut Components Supplement (18D80-18DE8) starts w/o a group header. It looks a bit odd. I had 'Components' which I think looked a bit better.
I added a subhead, but made it "Miscellaneous components" instead of
just "Components", because I think that contrasts better with the
following subheads that specify the number of strokes for each group of
components.
> - I had put the new 1F6D8 under the group header 'Map Symbols'. It is for sure a road sign (typically enclosed in a triangle). So  I would prefer to stay that way, and not just 'Miscellaneous Symbols', and as of now it creates three ugly <reserved> line under that new glyph.
Fixed.
> - By analogy/consistency with the group header for 1FAE6..1FAE7, the group header for 1FAEF should also be 'Emotion'.
O.k. Changed to that.
> - For 1FBFA (Alarm bell symbol), you don't mention it is not emoji which was part of the initial proposal text (I think). You do that for the other sets (such as 1CCFA..1CCFC)

This one I disagreed with. Unlike the other cases, which in the charts
do look like they could possibly be interpreted as stray emoji, this is
a single character that already has a subhead "Terminal graphic
character", and that, together with "SYMBOL" in the name seemed to
suffice for this one. Adding the notice in addition to the subhead in
this case seemed like it would actually detract by overkill. So I
applied the Joe Becker editorial principle. ;-)

* NamesList-17.0.0d8.txt

Yet another update, this one to roll in the change for [UTC-181-A147] (the
extended explanation about U+034F CGJ).

---------

Co-authored-by: Ken-Whistler <ken@unicode.org>