UCA 17.0 delta 1
From Ken on jan09 (before UTC-182 & 17 alpha):

The ordering was not complicated for most of
this repertoire, even though specific suggestions for collation were
missing from proposals in many instances. For full new scripts I added
in comments pointing to the relevant proposal documents, and for any
somewhat problematical cases, I added further notes in comments in the
input file.

The net net of this round was to increase the number of secondary
weights by 1 (263 --> 264), and the number of primary weights by 301
(28855 --> 29156).

I also went ahead and did the split for the implicit weighting of Tangut
ideographs versus Tangut components. The adds another couple
@implicitweights statements at the top of the file and bumps the initial
weight up for Nushu and Khitan Small Script.