Sifter updates for 17.0
From Ken on feb18:

The changes to the source code are minor -- just the usual updates for
new repertoire and date, plus some tweaks for the split for Tangut
implicit weighting.

I had to explicitly touch cttfooter.txt for the Tangut changes and for
the addition of Extension J, as well as the range for Extension C.

From Ken on feb19:

These were to fix the output for the CTT to include a time stamp on the
generation and to include the CTT table name *inside* the generated
file. The output now is just named "ctt.txt".

From Ken on apr24:

I've made the small updates to the sifter code and to the
CTT template text for the changes in CJK ranges.