From @Ken-Whistler:
Michel has provided what he intends as the final version of Unikemet.txt
for 17.0. I have pulled it into my framework and have done a quick check
that the changes look reasonable.

This includes normative changes to kEH_Desc field values that the UTC
will need to rubber-stamp this week, but also a bunch of changes to the
FVal and Func fields which show up in the names list.

Co-authored-by: Ken-Whistler <ken@unicode.org>
Co-authored-by: michelsu <67653968+michelsu@users.noreply.github.com>