Add five CJK confusables pairs from UTC 185 Action Items (#1236)
[185-A37] Action Item for Josh Hadley, Roozbeh Pournader, PAG: Add U+5BC9
and U+96BA as a new confusable pair, based on feedback [ID20250924032615] in
document L2/25-220 and Section 09 of document L2/25-231, for a future
version of the standard.

[185-A46] Action Item for Josh Hadley, Roozbeh Pournader, PAG: Add U+8EB1
and U+8EB2 as a new confusable pair, based on feedback [ID20251007094419] in
document L2/25-220 and Section 13 of document L2/25-231, for a future
version of the standard.

[185-A49] Action Item for Josh Hadley, Roozbeh Pournader, PAG: Add U+514C
and U+5151 as a new confusable pair, based on feedback [ID20251007134830] in
document L2/25-220 and Section 14 of document L2/25-231, for a future
version of the standard.

[185-A68] Action Item for Josh Hadley, Roozbeh Pournader, PAG: Add U+980B
and U+2EA07 as a new confusable pair, based on Recommendation M65.05 in
document IRG N2826 and Section 25 of document L2/25-231, for a future
version of the standard.

[185-A95] Action Item for Josh Hadley, Roozbeh Pournader, PAG: Add U+2EDB5
and U+32A8F as a new confusable pair, based on the kSpoofingVariant two
additions specified in the table in Section 38 of document L2/25-231, for a
future version of the standard.